backend/pyproject.toml
[project]
name = "agenikpredict-backend"
version = "0.1.0"
description = "AgenikPredict - Universal Collective Intelligence Engine, Predict Anything"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "AgenikPredict Team" }
]

dependencies = [
    # Core framework
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    # LLM
    "openai>=1.0.0",
    # Zep Cloud
    "zep-cloud==3.13.0",
    # File processing
    "PyMuPDF>=1.24.0",
    # Encoding detection
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "twelvedata>=1.2.28",
    "reticker>=3.1.6",
    "stripe>=14.4.1",
    "psycopg2-binary>=2.9.0",
    "gunicorn>=23.0.0",
    "boto3>=1.39.0",
]

[project.optional-dependencies]
simulation = [
    # OASIS social media simulation stack
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
